app/requirements.txt
cachetools==5.2.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
pydantic==1.10.2
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
python-dotenv==1.0.0
pylint==2.17.1